Create UInt64 from file size explicitly
Different libgit2 versions seem to treat this differently, so creating a
UInt64 explicitly fix this issue.

+++ Sources/SwiftGit2/Diffs.swift
@@ -57,7 +57,7 @@ public struct Diff {
 			self.oid = OID(diffFile.id)
 			let path = diffFile.path
 			self.path = path.map(String.init(cString:))!
-			self.size = diffFile.size
+			self.size = UInt64(diffFile.size)
 			self.flags = Flags(rawValue: diffFile.flags)
 		}
 	}
